requirements.txt
ply==3.11
lark-parser==0.12.0
pytest==7.4.3 